requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
boto3==1.35.96
botocore==1.35.96
branca==0.8.1
cachetools==5.5.0
catboost==1.2.7
certifi==2024.12.14
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
Django==5.1.4
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
earthengine-api==1.4.3
flatbuffers==24.12.23
fmodpy==1.7.4
folium==0.19.3
fonttools==4.55.3
future==1.0.0
gast==0.6.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.69.0
gunicorn==23.0.0
h5py==3.12.1
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mysqlclient==1.4.5
namex==0.0.8
numpy==1.26.4
nvidia-nccl-cu12==2.23.4
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
plotly==5.24.1
prettytable==3.12.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycup==0.1.7.2
pyDOE==0.3.8
Pygments==2.19.0
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
pyswarms==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
scikit-learn==1.2.2
scipy==1.14.1
six==1.17.0
spotpy==1.6.2
sqlparse==0.5.3
statsmodels==0.14.4
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.8.2
wrapt==1.17.0
xgboost==1.7.3
xyzservices==2024.9.0
